+++ src/components/sidebar/LogsCtrl.tsx
@@ -67,9 +67,11 @@ export default defineComponent({
           v-model={logFilter.value}
           beforeClose={true}
           class="flex-1"
-          placeholder={t('search')}
+          placeholder={`${t('search')} | Regex`}
           clearable={true}
           menus={logSearchHistory.value}
+          menusDeleteable={true}
+          onUpdate:menus={(val) => (logSearchHistory.value = val)}
         />
       )