+++ src/i18n/en.ts
@@ -138,6 +138,8 @@ const en = {
   manual: 'Manual',
   tableWidthMode: 'Table Width Mode',
   testFinishedTip: '{number} Test Finished',
+  testFinishedResultTip: 'Test Finished, {success} Success, {failed} Timeout',
+  testFailedTip: 'Latency Test Timeout',
   updateFinishedTip: '{number} Update Finished',
   independentLatencyTest: 'Independent Latency Test',
   independentLatencyTestTip:
@@ -183,6 +185,15 @@ const en = {
   importFromFile: 'Import From File',
   importFromUrl: 'Import From URL',
   sync: 'Sync',
+  upgradeSuccess: 'Upgrade Success',
+  numberOfChartsInSidebar: 'Number of Charts in Sidebar',
+  displayWeights: 'Display Weights',
+  flushWeights: 'Flush Weights',
+  flushSmartWeights: 'Flush Smart Weights',
+  IPInfoAPI: 'IP Info API',
+  IPInfoAPITip:
+    "This API will be used for IP checks in global node IP information queries, IP geolocation queries in connection details, and IP geolocation queries in zashboard's DNS query feature.",
+  general: 'General',
 }
 
 export type LANG_MESSAGE = typeof en