+++ src/i18n/zh.ts
@@ -139,6 +139,8 @@ const zh: LANG_MESSAGE = {
   manual: '手动',
   tableWidthMode: '表格宽度模式',
   testFinishedTip: '{number} 测试完成',
+  testFinishedResultTip: '测试完成，{success} 测试成功，{failed} 测试超时',
+  testFailedTip: '测速超时',
   updateFinishedTip: '{number} 更新完成',
   independentLatencyTest: '独立延迟测试',
   independentLatencyTestTip:
@@ -182,6 +184,15 @@ const zh: LANG_MESSAGE = {
   importFromFile: '从文件导入',
   importFromUrl: '从 URL 导入',
   sync: '同步',
+  upgradeSuccess: '更新成功',
+  numberOfChartsInSidebar: '侧边栏图表数量',
+  displayWeights: '显示权重',
+  flushWeights: '清空权重',
+  flushSmartWeights: '清空Smart权重',
+  IPInfoAPI: 'IP信息API',
+  IPInfoAPITip:
+    '此API会用于IP检查中全球节点IP信息查询、连接详情中的IP地理信息查询、面板DNS查询中的IP地理信息查询。',
+  general: '通用',
 }
 
 export default zh